requirements.txt
# Django and core dependencies
Django==5.1.4
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# GeoDjango dependencies
GDAL==3.8.2
Fiona==1.9.5
Shapely==2.0.2

# REST Framework
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# CORS handling
django-cors-headers==4.3.1

# Static files handling
whitenoise==6.6.0

# Form utilities
django-widget-tweaks==1.5.0

# Production server
gunicorn==21.2.0

# Environment variables
python-decouple==3.8

# Utilities
Pillow==10.1.0
pytz==2023.3
